+++ oap-server/server-library/library-async-profiler-jfr-parser/src/main/java/one/jfr/CallStack.java
@@ -3,10 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.apache.skywalking.oap.server.library.jfr.parser.convert;
+package one.jfr;
+
+import lombok.Getter;
 
 import java.util.Arrays;
 
+/**
+ * This class is placed in the one.jfr package because some classes and fields of the async-profiler-converter package
+ * can only be accessed under the same package name, and what we want to do is to expand it, so here we choose to create
+ * a class with the same package name for extension.
+ */
+@Getter
 public class CallStack {
     String[] names = new String[16];
     byte[] types = new byte[16];
