+++ oap-server/server-library/library-async-profiler-jfr-parser/src/main/java/org/apache/skywalking/oap/server/library/jfr/parser/JFRMergeBuilder.java
@@ -16,18 +16,19 @@
  *
  */
 
-package org.apache.skywalking.oap.server.library.jfr.parser.convert;
+package org.apache.skywalking.oap.server.library.jfr.parser;
 
 import java.util.List;
 
-import static org.apache.skywalking.oap.server.library.jfr.parser.convert.Frame.TYPE_INTERPRETED;
-import static org.apache.skywalking.oap.server.library.jfr.parser.convert.Frame.TYPE_NATIVE;
+import static org.apache.skywalking.oap.server.library.jfr.parser.Frame.TYPE_INTERPRETED;
+import static org.apache.skywalking.oap.server.library.jfr.parser.Frame.TYPE_NATIVE;
 
-public class JfrMergeBuilder {
+
+public class JFRMergeBuilder {
     private final Index<String> cpool = new Index<>(String.class, "");
     private final Frame root = new Frame(0, TYPE_NATIVE);
 
-    public JfrMergeBuilder merge(List<FrameTree> trees) {
+    public JFRMergeBuilder merge(List<FrameTree> trees) {
         if (trees == null || trees.isEmpty()) {
             return this;
         }
@@ -37,7 +38,7 @@ public class JfrMergeBuilder {
         return this;
     }
 
-    public JfrMergeBuilder merge(FrameTree tree) {
+    public JFRMergeBuilder merge(FrameTree tree) {
         merge0(root, tree);
         return this;
     }
